Step4 - 로또(수동) (#641)
* feat(manual): 로또(수동)기능 추가

- 로또 번호 생성기 -> 로또 일급 컬렉션 생성기로 변경
- 수동 로또 목록 생성기 구현
- 수동 로또 목록 생성기 테스트 케이스 작성
- 로또 일급 컬렉션 델리게이션 적용
- 요청정보 그룹핑
- 로또 UI Layer(Input) 프록시 패턴 적용
- 돈 객체 타입 변경 (Int -> Long)

* docs: update README.md

- 코드 klint format 적용

* refactor: 피드백 반영

- 연산자 오버로딩 이용한 일급 컬렉션 병합 로직 수정
- custom getter() 초기화 제거
- request 필드 타입 변경
- try-catch -> runCatching 방식 변경
- 로또 생성 로직 수정
- 구조분해 방식 롤백

* refactor: 피드백 반영

- unused field 제거

---------

+++ src/test/kotlin/lotto/domain/MoneyTest.kt
@@ -6,7 +6,7 @@ import io.kotest.matchers.shouldBe
 
 class MoneyTest : FunSpec({
     test("유효한 자연수를 전달하면 돈 객체가 생성된다.") {
-        listOf(0, 1, 10, 100, 1000, 10000, 10000, Int.MAX_VALUE).forEach {
+        listOf(0L, 1L, 10L, 100L, 1000L, 10000L, 10000L, Long.MAX_VALUE).forEach {
             val actual = Money(it)
 
             actual.value shouldBe it
@@ -22,7 +22,7 @@ class MoneyTest : FunSpec({
     }
 
     test("유효하지 않은 인자를 전달하면 예외를 던진다.") {
-        listOf(-1, -100, Int.MIN_VALUE).forEach {
+        listOf(-1L, -100L, Long.MIN_VALUE).forEach {
             shouldThrow<IllegalArgumentException> { Money(it) }
         }
